insert_youbike_data/requirements.txt
psycopg2-binary
requests
pydantic
python-dotenv
python-crontab